% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{blockDevices}
\alias{blockDevices}
\title{Describe one or more block devices (EBS) to create, format, and mount on an
EC2 instance}
\usage{
blockDevices(
  deviceName = "/dev/xvdf",
  volumeSize = 8,
  deleteOnTermination = TRUE,
  volumeType = "gp2",
  encrypted = FALSE,
  fsType = "ext4",
  mountDir = "/data"
)
}
\arguments{
\item{deviceName}{A character vector giving the names of one or more block
devices to create on instance launch.  The length of \code{mountDir} must
match the length of this vector.}

\item{volumeSize}{A numeric vector giving the size in GB of the volume(s) to
create.}

\item{deleteOnTermination}{TRUE or FALSE to determine if the volume is
deleted when the associated EC2 instance is terminated. \bold{WARNING #1}:
Even if set to TRUE, the volume will not be deleted if the attached EC2
instance is stopped instead of terminated.  \bold{WARNING #2}:  If you set
this to FALSE than you will continue to incur EBS charges for data on this
volume even after the instance has been terminated.}

\item{volumeType}{Type of EBS volume: "standard"|"io1"|"gp2"|"sc1"|"st1".}

\item{encrypted}{TRUE or FALSE whether or not the drive should be encrypted.}

\item{fsType}{The type of filesystem to use when formatting block devices.
This can be any valid argument to 'mkfs -t' on your AMI.}

\item{mountDir}{The path on the instance file system where the device(s)
should be mounted after being formatted.  The length of this vector must
math the length of \code{deviceName}.}
}
\value{
A list of block device mappings that can be used with
  \code{\link{launchInstance}}.
}
\description{
The arguments to this function are vectorized, so you can use it to describe
one or more block devices to create and mount.  Pay particularly close
attention to the \code{deleteOnTermination} argument!  If set incorrectly you
may unexpectedly lose data or incur EBS charges.
}
\details{
Note the design of this package assumes you are using attached block devices
for "scratch" drives during a data analysis run that are freshly provisioned
at the start and discarded at the end.  If this doesn't fit your use-case
you'll probably want to manually launch and configure your instances and
attached volumns.
}
