% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{getInstanceDescription}
\alias{getInstanceDescription}
\title{Gather descriptive information about a running instance}
\usage{
getInstanceDescription(instanceId, throttle = 1)
}
\arguments{
\item{instanceId}{The InstanceID returned or displayed through whatever
mechanism used to launch the EC2 instance}

\item{throttle}{The amount of time to wait (in seconds) before repeating
requests to the AWS API.  The account-wide limit is 10,000 r/s, so this
default is extremely conservative.}
}
\value{
An \code{instance} object used by the other functions in this package
}
\description{
This function checks to make sure a launched instance is Running and both the
System and Instance reachability tests have passed, then returns an instance
description used by the other functions in this package.  This function is
useful if you want to launch instances yourself using the Web console, AWS
CLI, or paws SDK instead of using \code{\link{launchInstance}}.
}
