% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{launchInstance}
\alias{launchInstance}
\title{Launch and EC2 instance}
\usage{
launchInstance(imageID, keyName, securityGroups,
  instanceType = "t2.micro", tags = NULL, blockDeviceMapping = NULL)
}
\arguments{
\item{imageID}{The image-id for the AMI to use to launch the instance.}

\item{keyName}{The key-pair on your AWS account to use.  This pair should be
associated with one of the default public keys that `ssh` will try to use
when connecting to your instance.}

\item{securityGroups}{The EC2 security group names to use.  This function
launches your instance with the EC2-Classic scheme, not using VPC.}

\item{instanceType}{The type of instance to launch.}

\item{tags}{Tags to apply to your instance and its resources. This should
either be NULL or be a normal R list with named elements; names will be tag
Keys and elements their Values.}

\item{blockDeviceMapping}{Additional block devices to create and mount. This
should either be NULL or an object returned by the
\code{\link{blockDeviceMapping}} function.}
}
\description{
This wrapper function launches an EC2 instance from an AMI image ID, monitors
instance initialization and startup, and handles mounting any additional
block devices at the specified locations.  It returns an \code{instance} object
which is used by other functions in this package to communicate with the
instance.
}
\details{
This function is designed to be as simple as possible and does not support
the vast majority of configuration options available when launching EC2
instances.  If it doesn't meet your needs see the \code{paws} package for a full
featured EC2 API.  You can pass the return value from the \code{paws.compute}
\code{run_instances} function to the \code{\link{parseLaunchResponse}} to get an
\code{instance} object to use with the other functions in this package.
}
