% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{launchJobs}
\alias{launchJobs}
\title{Start jobs by executing a launch script on an EC2 instance}
\usage{
launchJobs(instance, scripts, otherFiles = NULL, permissions = "ug+x")
}
\arguments{
\item{instance}{An \code{instance} objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}

\item{otherFiles}{NULL or a vector containing full paths to other files to be
copied to the instance before the launch script is run.}

\item{permissions}{Specification for how the executable bit should be set
after scripts are copied to the instance.  By default this marks the scrips
as executable by both the owning user and group.}
}
\value{
The \code{instance} object.
}
\description{
This function copies your individual job scripts and the launch script to an
EC2 instance, flags them as executable, and then runs them.
}
