% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell-scripts.R
\name{makeJobScripts}
\alias{makeJobScripts}
\title{Generate a set of shell scripts to to launch individual jobs}
\usage{
makeJobScripts(template, processes = 16,
  scriptPath = "rendered-scripts", ...)
}
\arguments{
\item{template}{Path to the script whisker template.}

\item{processes}{How many processes launched jobs are sharing  This makes a
`threads` variable available during template rendering, which doles out an
even number of threads to each job based on the total process count.}

\item{scriptPath}{Path to the folder where rendered scripts should be saved.}

\item{...}{A variable number of other arguments that are available as
variable bindings.  Values not of length 1 should all be of the same length
(the number of job scripts being created).  Must contain at least a `file`
argument which lists the output file names for rendered scripts.}
}
\value{
Returns a character vector containing the names of the job script
  files created.
}
\description{
This function uses a whisker (aka Mustache) template file to create one or
more job scripts.
}
