% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell-scripts.R
\name{makeLaunchScript}
\alias{makeLaunchScript}
\title{Generate a script to launch a batch of jobs}
\usage{
makeLaunchScript(scripts, launchScript = "launch.sh",
  scriptPath = "rendered-scripts", logPath = "logs", prelaunch = "",
  postlaunch = "wait; sudo shutdown now")
}
\arguments{
\item{scripts}{File names of job scripts to add to launch.  Should be located
in `scriptPath`.}

\item{launchScript}{The name of the launch script file to be created.}

\item{scriptPath}{Path to the folder where jobs scripts should be saved and
launch script will be created.}

\item{logPath}{Directory where logs recording stdout and stderr streams for
each job should be saved.}

\item{prelaunch}{Commands to run before the jobs are launched.}

\item{postlaunch}{Commands to run after all jobs have finished.  By default
it shutsdown the instance when all of the jobs are done (assuming the user
running the launch script has sudoer priveledges).  Set to an empty string
to override this behavior.}
}
\description{
The launch script starts jobs as background processes, records process ids in
a `jobs.pid` file, and by default `waits` for completion of all of the jobs
and then shuts down the instance.  It also redirects both the stdout and
stderr streams into log files.
}
