% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell-scripts.R
\name{makeQueueScript}
\alias{makeQueueScript}
\title{Generate a queue script to launch a batch of jobs}
\usage{
makeQueueScript(scripts, queueScript = "queue.sh",
  scriptPath = "rendered-scripts", otherPaths = "logs",
  prequeue = "", postqueue = "wait; sudo shutdown now")
}
\arguments{
\item{scripts}{File names of job scripts to add to the queue.  Should be
located in `scriptPath`.}

\item{queueScript}{The name of the queue script file to be created.}

\item{scriptPath}{Path to the folder where jobs scripts should be saved and
queue script will be created.}

\item{otherPaths}{Any directories that need to be created before the jobs are
run.}

\item{prequeue}{Commands to run before the jobs are launched.}

\item{postqueue}{Commands to run after all jobs have finished.  By default it
shutsdown the instance when all of the jobs are done (if the user running
the queue script has sudo priveledges).  Set to an empty string to override
this behavior.}
}
\description{
The queue script launches jobs as background processes, records process ids
in a `jobs.pid` file, and by default `waits` for completion of all of the
jobs and then shuts down the instance.  It also redirects both the stdout and
stderr streams into log files located in `logs/`.
}
