% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-monitor.R
\name{monitor}
\alias{monitor}
\title{Launch instance monitoring app}
\usage{
monitor(instances, updateFrequency = 2)
}
\arguments{
\item{instances}{Either an \code{instance} objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}} or a
list containing one or more \code{instance} objects.}

\item{updateFrequency}{Status update frequency (in seconds).  Note: this sets
the delay between update attempts; server lag may increase the effective
update rate.}
}
\description{
Launches a shiny app for real-time monitoring of one or more EC2 instances
and the jobs running on them.
}
\details{
This feature requires both the \code{top} and \code{mpstat} programs be
installed on the instance.
}
