% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash-scripts.R
\name{renderJobScripts}
\alias{renderJobScripts}
\title{Render job scripts from template files}
\usage{
renderJobScripts(template, processes = 16,
  queueScript = paste("queue-", basename(template), sep = ""),
  scriptPath = "rendered-scripts", otherPaths = "logs",
  prequeue = "",
  postqueue = "wait\\necho \\"All jobs have finished!\\"", ...)
}
\arguments{
\item{template}{File path to the script template}

\item{processes}{How many processes queued jobs are sharing.  This makes a
`threads` variable available during template rendering.}

\item{queueScript}{If not FALSE than the name of the queue script file.}

\item{scriptPath}{Path to the folder where rendered scripts should be saved.}

\item{otherPaths}{Other directories that need to be created before the jobs
are run.}

\item{prequeue}{Commands to run before the jobs are launched.}

\item{postqueue}{Commands to run after all jobs have finished.}

\item{...}{A variable number of other arguments that are available as
variable bindings.  Values not of length 1 should all be of the same
length.  Must contain at least `file` entry, listing the file names for
rendered scripts.}
}
\description{
This function uses a whisker (aka Mustache) template file to creates one or
more job queue scripts.  It also optionally can create a bash script that
launches all of the jobs.
}
