% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{startInstance}
\alias{startInstance}
\title{Starts a stopped instance}
\usage{
startInstance(instanceId, throttle = 1, username = "ec2-user", sshTimeout = 10)
}
\arguments{
\item{instanceId}{The Instance ID for the stopped instance to start.}

\item{throttle}{The amount of time to wait (in seconds) before repeating
requests to the AWS API.  The account-wide limit is 10,000 r/s, so this
default is extremely conservative.}

\item{username}{The administrator username to use when logging in over ssh.
The user must have sudo privileges. Default is set for Amazon Linux AMIs.}

\item{sshTimeout}{How many times to re-test connecting over ssh after
recieving a timeout or connection refused error.  This function waits the
duraction of 'throttle' between attempts.}
}
\value{
An \code{instance} object that can be used with the other functions
  in this package.
}
\description{
Starts a stopped instance and waits for the instance to start running and for
ssh access to be available.
}
