% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{stopInstance}
\alias{stopInstance}
\title{Stop a running instance}
\usage{
stopInstance(instance)
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}
}
\value{
The Instance ID for the stopped instance.
}
\description{
Stop a running instance
}
