% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{terminateInstance}
\alias{terminateInstance}
\title{Title}
\usage{
terminateInstance(instance)
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}
}
\description{
Title
}
