% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{waitForJobs}
\alias{waitForJobs}
\title{Wait for jobs on an instance to finish}
\usage{
waitForJobs(instance, wait = 10)
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}

\item{wait}{The time (in seconds) to wait between checking on the state of jobs
running on the instance.}
}
\value{
Invisibly returns the \code{instance} object.
}
\description{
This function blocks the current thread of execution until all of the jobs on
the given instance have finished running.
}
