% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{getStatus}
\alias{getStatus}
\title{Get the status of jobs launched on an EC2 instance.}
\usage{
getStatus(instance)
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}
}
\description{
Get the status of jobs launched on an EC2 instance.
}
