% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{getStatus}
\alias{getStatus}
\title{Get the status of an EC2 instance.}
\usage{
getStatus(instance, session = NULL)
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}

\item{session}{Optionally provide an existing ssh session to the instance
created with ssh::ssh_connect.}
}
\value{
A list containing two named elemements: \code{jobs} a data.frame
  giving the status of all jobs launched on the instance and \code{system} a
  list containing system summary information.
}
\description{
This function requires the \code{top} and \code{mpstat} programs on the
instance.
}
