% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{launchJobs}
\alias{launchJobs}
\title{Start jobs by executing a launch script on an EC2 instance}
\usage{
launchJobs(instance, launchScript = "launch.sh",
  scriptPath = "rendered-scripts")
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}

\item{launchScript}{The name of the launch script file created by
\code{\link{makeLaunchScript}}.}

\item{scriptPath}{Path to the folder where jobs scripts and the launch script
were created with \code{\link{makeJobScripts}} and
\code{\link{makeLaunchScript}}.}
}
\description{
This function copies your individual job scripts and the launch script to an
EC2 instance, flags them as executable, and then runs them.
}
