% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell-scripts.R
\name{makeJobScripts}
\alias{makeJobScripts}
\title{Generate a set of shell scripts to launch individual jobs}
\usage{
makeJobScripts(
  template,
  jobScripts,
  processes = 1,
  scriptPath = "rendered-scripts",
  ...
)
}
\arguments{
\item{template}{Path to the script whisker template.}

\item{jobScripts}{The file names for the job scripts to create through
templating.  Should match the length of the longest variable argument.}

\item{processes}{The number of processes launched jobs are sharing  This
makes a `threads` variable available during template rendering, which doles
out an even number of threads to each job based on the total process count.}

\item{scriptPath}{Path to the local folder where rendered scripts should be
saved.}

\item{...}{A variable number of other arguments that are available as
variable bindings when the template is rendered.  Values not of length 1
should all be of the same length (the number of job scripts being created).}
}
\value{
A List with one named element (\code{jobs}) containing a vector of
  job script file names.  Passed to \code{\link{makeLaunchScript}}
}
\description{
This function uses a whisker (aka Mustache) template file to create one or
more job scripts.
}
