% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell-scripts.R
\name{makeLaunchScript}
\alias{makeLaunchScript}
\title{Generate a script to launch a batch of jobs}
\usage{
makeLaunchScript(
  scripts,
  launchScript = "launch.sh",
  workingPath = "",
  logs = "logs",
  username = "ec2-user",
  prelaunch = "",
  postlaunch = "wait; sudo shutdown now"
)
}
\arguments{
\item{scripts}{List with job script file names returned by
\code{\link{makeJobScripts}}.}

\item{launchScript}{The name of the launch script file to be created.}

\item{workingPath}{Path to the working directory on remote EC2 instances
where scripts should be copied and executed. Default is /home/\code{username}/.}

\item{logs}{Folder where logs recording stdout and stderr streams for each
job should be saved.  This is created as a subdirectory of
\code{workingPath}.}

\item{username}{The username being used on the instance.}

\item{prelaunch}{Commands to run before the jobs are launched.}

\item{postlaunch}{Commands to run after all jobs have finished.  The default
shuts down the instance after all of the jobs have finished running.  Set
to an empty string to override this behavior.}
}
\value{
Named list with job and launch script file names.  Passed to
  \code{\link{launchJobs}}.
}
\description{
The launch script starts jobs as background processes, records process ids in
a `jobs.pid` file, and by default `waits` for completion of all of the jobs
and then shuts down the instance.  It also redirects both the stdout and
stderr streams into log files.
}
