% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{terminateInstance}
\alias{terminateInstance}
\title{Terminate an instance}
\usage{
terminateInstance(instance)
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}
}
\value{
Invisibly returns the \code{instance} object.
}
\description{
Terminates the instance. \bold{WARNING}: any block devices setup with
\code{deleteOnTermination = TRUE} will be deleted and ALL data LOST!  A
termianted instance cannot be restarted.
}
