% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{terminateInstance}
\alias{terminateInstance}
\title{Terminate an instance}
\usage{
terminateInstance(instance)
}
\arguments{
\item{instance}{An instance objected created with
\code{\link{launchInstance}} or \code{\link{getInstanceDescription}}.}
}
\description{
Terminates the instance.  NOTE: any block devices setup with
\code{deleteOnTermination = TRUE} will be deleted and ALL data LOST!
}
\details{
This package doesn't offer an API for stopping and resuming instances, only
terminating them.  If you want to manually stop and start instances, you can
get a new \code{instance} object to use with the other functions in this
package by passing the instanceId to \code{\link{getInstanceDescription}}
function.
}
